Split out merge transition validator
