+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/forkchoice/ForkChoice.java
@@ -188,7 +188,7 @@ public class ForkChoice {
         blockSlotState.get().getSlot());
 
     final ForkChoicePayloadExecutor payloadExecutor =
-        new ForkChoicePayloadExecutor(spec, recentChainData, block, executionEngine);
+        ForkChoicePayloadExecutor.create(spec, recentChainData, block, executionEngine);
     final ForkChoiceUtil forkChoiceUtil = spec.atSlot(block.getSlot()).getForkChoiceUtil();
     final BlockImportResult preconditionCheckResult =
         forkChoiceUtil.checkOnBlockConditions(
