+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/forkchoice/ForkChoicePayloadExecutorTest.java
@@ -22,28 +22,20 @@ import static org.mockito.Mockito.when;
 import static tech.pegasys.teku.spec.executionengine.PayloadStatus.VALID;
 
 import java.util.Optional;
-import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import tech.pegasys.teku.core.ChainBuilder.BlockOptions;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.TestSpecFactory;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeader;
-import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.bellatrix.BeaconStateBellatrix;
 import tech.pegasys.teku.spec.executionengine.ExecutionEngineChannel;
 import tech.pegasys.teku.spec.executionengine.PayloadStatus;
 import tech.pegasys.teku.spec.logic.common.block.AbstractBlockProcessor;
-import tech.pegasys.teku.spec.logic.versions.bellatrix.helpers.BellatrixTransitionHelpers;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionsBellatrix;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
-import tech.pegasys.teku.storage.storageSystem.InMemoryStorageSystemBuilder;
-import tech.pegasys.teku.storage.storageSystem.StorageSystem;
 
 class ForkChoicePayloadExecutorTest {
 
@@ -57,7 +49,8 @@ class ForkChoicePayloadExecutorTest {
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
   private final SafeFuture<PayloadStatus> executionResult = new SafeFuture<>();
   private final ExecutionEngineChannel executionEngine = mock(ExecutionEngineChannel.class);
-  private final StorageSystem storageSystem = InMemoryStorageSystemBuilder.buildDefault(spec);
+  private final MergeTransitionBlockValidator transitionValidator =
+      mock(MergeTransitionBlockValidator.class);
   private final ExecutionPayloadHeader payloadHeader =
       dataStructureUtil.randomExecutionPayloadHeader();
   private final ExecutionPayload payload = dataStructureUtil.randomExecutionPayload();
@@ -74,9 +67,7 @@ class ForkChoicePayloadExecutorTest {
 
   @BeforeEach
   void setUp() {
-    storageSystem.chainUpdater().initializeGenesis(false);
     when(executionEngine.newPayload(any())).thenReturn(executionResult);
-    when(executionEngine.getPowBlock(any())).thenReturn(new SafeFuture<>());
   }
 
   @Test
@@ -96,14 +87,6 @@ class ForkChoicePayloadExecutorTest {
     assertThat(payloadExecutor.getExecutionResult()).isCompletedWithValue(PayloadStatus.VALID);
   }
 
-  /**
-   * The details of how we validate the merge block are all handled by {@link
-   * BellatrixTransitionHelpers} so we don't want to retest all that here. Instead, we check that
-   * those extra checks started (the call to {@link ExecutionEngineChannel#getPowBlock(Bytes32)}).
-   *
-   * <p>Since the future we return from getPowBlock is never completed we never complete the
-   * validations which saves us a bunch of mocking.
-   */
   @Test
   void optimisticallyExecute_shouldValidateMergeBlockWhenThisIsTheMergeBlock() {
     when(executionEngine.newPayload(payload)).thenReturn(SafeFuture.completedFuture(VALID));
@@ -113,7 +96,7 @@ class ForkChoicePayloadExecutorTest {
 
     // Should execute first and then begin validation of the transition block conditions.
     verify(executionEngine).newPayload(payload);
-    verify(executionEngine).getPowBlock(payload.getParentHash());
+    verify(transitionValidator).verifyTransitionBlock(defaultPayloadHeader, payload);
     assertThat(result).isTrue();
   }
 
@@ -124,7 +107,7 @@ class ForkChoicePayloadExecutorTest {
     final boolean execution = payloadExecutor.optimisticallyExecute(defaultPayloadHeader, payload);
 
     // Should not attempt to validate transition conditions because execute payload failed
-    verify(executionEngine, never()).getPowBlock(payload.getParentHash());
+    verify(transitionValidator, never()).verifyTransitionBlock(defaultPayloadHeader, payload);
     verify(executionEngine).newPayload(payload);
     assertThat(execution).isTrue();
     assertThat(payloadExecutor.getExecutionResult())
@@ -135,12 +118,12 @@ class ForkChoicePayloadExecutorTest {
   void
       optimisticallyExecute_shouldReturnFailedExecutionOnMergeBlockWhenELGoesOfflineAfterExecution() {
     when(executionEngine.newPayload(payload)).thenReturn(SafeFuture.completedFuture(VALID));
-    when(executionEngine.getPowBlock(payload.getParentHash()))
+    when(transitionValidator.verifyTransitionBlock(defaultPayloadHeader, payload))
         .thenReturn(SafeFuture.failedFuture(new Error()));
     final ForkChoicePayloadExecutor payloadExecutor = createPayloadExecutor();
     final boolean execution = payloadExecutor.optimisticallyExecute(defaultPayloadHeader, payload);
 
-    verify(executionEngine).getPowBlock(payload.getParentHash());
+    verify(transitionValidator).verifyTransitionBlock(defaultPayloadHeader, payload);
     verify(executionEngine).newPayload(payload);
     assertThat(execution).isTrue();
     assertThat(payloadExecutor.getExecutionResult())
@@ -148,7 +131,7 @@ class ForkChoicePayloadExecutorTest {
   }
 
   @Test
-  void optimisticallyExecute_shouldReturnNotVerifyTransitionIfExecutePayloadIsInvalid() {
+  void optimisticallyExecute_shouldNotVerifyTransitionIfExecutePayloadIsInvalid() {
     final PayloadStatus expectedResult =
         PayloadStatus.invalid(Optional.empty(), Optional.of("Nope"));
     when(executionEngine.newPayload(payload))
@@ -157,7 +140,7 @@ class ForkChoicePayloadExecutorTest {
     final boolean execution = payloadExecutor.optimisticallyExecute(defaultPayloadHeader, payload);
 
     verify(executionEngine).newPayload(payload);
-    verify(executionEngine, never()).getPowBlock(payload.getParentHash());
+    verify(transitionValidator, never()).verifyTransitionBlock(defaultPayloadHeader, payload);
     assertThat(execution).isTrue();
     assertThat(payloadExecutor.getExecutionResult()).isCompletedWithValue(expectedResult);
   }
@@ -172,6 +155,8 @@ class ForkChoicePayloadExecutorTest {
 
   @Test
   void shouldReturnExecutionResultWhenExecuted() {
+    when(transitionValidator.verifyTransitionBlock(payloadHeader, payload))
+        .thenReturn(SafeFuture.completedFuture(VALID));
     final ForkChoicePayloadExecutor payloadExecutor = createPayloadExecutor();
     payloadExecutor.optimisticallyExecute(payloadHeader, payload);
 
@@ -183,109 +168,7 @@ class ForkChoicePayloadExecutorTest {
     assertThat(result).isCompletedWithValue(VALID);
   }
 
-  @Test
-  void shouldVerifyNonFinalizedAncestorTransitionBlock() {
-    final SignedBlockAndState transitionBlock = generateNonfinalizedTransition();
-    final SignedBlockAndState chainHead = storageSystem.chainBuilder().getLatestBlockAndState();
-    final SignedBlockAndState blockToVerify = storageSystem.chainBuilder().generateNextBlock();
-
-    final ForkChoicePayloadExecutor payloadExecutor =
-        createPayloadExecutor(blockToVerify.getBlock());
-    final ExecutionPayload newExecutionPayload = getExecutionPayload(blockToVerify);
-    when(executionEngine.newPayload(newExecutionPayload))
-        .thenReturn(SafeFuture.completedFuture(VALID));
-
-    assertThat(
-            storageSystem
-                .recentChainData()
-                .getForkChoiceStrategy()
-                .orElseThrow()
-                .getOptimisticallySyncedTransitionBlockRoot(chainHead.getRoot()))
-        .isPresent();
-
-    payloadExecutor.optimisticallyExecute(
-        chainHead.getState().toVersionBellatrix().orElseThrow().getLatestExecutionPayloadHeader(),
-        newExecutionPayload);
-
-    verify(executionEngine).newPayload(newExecutionPayload);
-    verify(executionEngine).getPowBlock(getExecutionPayload(transitionBlock).getParentHash());
-    assertThat(payloadExecutor.getExecutionResult()).isNotCompleted();
-  }
-
-  @Test
-  void shouldVerifyFinalizedAncestorTransitionBlock() {
-    final SignedBlockAndState transitionBlock = generateFinalizedTransition();
-    final BeaconStateBellatrix chainHeadState =
-        storageSystem
-            .chainBuilder()
-            .getLatestBlockAndState()
-            .getState()
-            .toVersionBellatrix()
-            .orElseThrow();
-
-    final SignedBlockAndState blockToVerify = storageSystem.chainBuilder().generateNextBlock();
-
-    final ForkChoicePayloadExecutor payloadExecutor =
-        createPayloadExecutor(blockToVerify.getBlock());
-    final ExecutionPayload newExecutionPayload = getExecutionPayload(blockToVerify);
-    when(executionEngine.newPayload(newExecutionPayload))
-        .thenReturn(SafeFuture.completedFuture(VALID));
-
-    assertThat(storageSystem.recentChainData().getStore().getFinalizedOptimisticTransitionPayload())
-        .isPresent();
-
-    payloadExecutor.optimisticallyExecute(
-        chainHeadState.getLatestExecutionPayloadHeader(), newExecutionPayload);
-
-    verify(executionEngine).newPayload(newExecutionPayload);
-    verify(executionEngine).getPowBlock(getExecutionPayload(transitionBlock).getParentHash());
-    assertThat(payloadExecutor.getExecutionResult()).isNotCompleted();
-  }
-
-  private SignedBlockAndState generateNonfinalizedTransition() {
-    storageSystem
-        .chainBuilder()
-        .generateBlocksUpToSlot(5)
-        .forEach(storageSystem.chainUpdater()::saveOptimisticBlock);
-    final Bytes32 terminalBlockHash = dataStructureUtil.randomBytes32();
-    final SignedBlockAndState transitionBlock =
-        storageSystem
-            .chainBuilder()
-            .generateBlockAtSlot(6, BlockOptions.create().setTerminalBlockHash(terminalBlockHash));
-    storageSystem.chainUpdater().saveOptimisticBlock(transitionBlock);
-
-    storageSystem
-        .chainBuilder()
-        .generateBlocksUpToSlot(39)
-        .forEach(storageSystem.chainUpdater()::saveOptimisticBlock);
-    final SignedBlockAndState chainHead = storageSystem.chainBuilder().generateBlockAtSlot(40);
-    storageSystem.chainUpdater().saveOptimisticBlock(chainHead);
-    return transitionBlock;
-  }
-
-  private SignedBlockAndState generateFinalizedTransition() {
-    final SignedBlockAndState transitionBlock = generateNonfinalizedTransition();
-
-    // Finalize between transition block and chain head
-    storageSystem.chainUpdater().finalizeEpoch(1);
-    return transitionBlock;
-  }
-
-  private ExecutionPayload getExecutionPayload(final SignedBlockAndState blockToVerify) {
-    return blockToVerify
-        .getBlock()
-        .getMessage()
-        .getBody()
-        .getOptionalExecutionPayload()
-        .orElseThrow();
-  }
-
   private ForkChoicePayloadExecutor createPayloadExecutor() {
-    return createPayloadExecutor(storageSystem.chainBuilder().generateNextBlock().getBlock());
-  }
-
-  private ForkChoicePayloadExecutor createPayloadExecutor(final SignedBeaconBlock block) {
-    return new ForkChoicePayloadExecutor(
-        spec, storageSystem.recentChainData(), block, executionEngine);
+    return new ForkChoicePayloadExecutor(executionEngine, transitionValidator);
   }
 }
